requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
google-genai>=1.0.0
python-dotenv==1.0.1
pydantic>=2.0.0